initial commit
This project was originally completed and pushed to my old (but still active) github account at https://github.com/obycodes/Bookbot, as a requirement to demonstrate knowledge of beginner and intermediate python as an aspiring and self-taught software engineer. Finally, I am doing this in an effort to consolidate some of my relevant and previous projects in one github account.